README: Mention "setup, clone, serve" at the top
This makes it easier for the reader to understand, and prevents the need
for the user to remember to update this text in their own repo, since
they should replace everything in the top section anyway.

Also renames one of the sections to replace "Guides Template" with
"site", to make the instructions more generic.